requirements.txt
flask
opencv-python
numpy
python-dotenv
groq
